Finish tcs

+++ 007.Reverse_Integer/solution.h
@@ -1,12 +1,14 @@
+#include <climits>
+
 class Solution {
 public:
     int reverse(int x) {
         long result = 0;
-        while (x){
+        while (x) {
             result = result * 10 + x % 10;
-            if (result<INT_MIN || result>INT_MAX) return 0;
+            if (result < INT_MIN || result > INT_MAX) return 0;
             x /= 10;
         }
-        return result;
+        return static_cast<int>(result);
     }
 };
